Polyfill Promise.withResolvers().
This hopefully improves compatibility with old Safari versions and Hermes (React Native).

Unfortunately it didn't seem easy to extend the tests to cover React Native. There is a vitest-react-native package, but it looks little-used and unmaintained, so I didn't want to install it.

Might help with #91, though I won't declare that one "fixed" until we actually have tests proving it.